Fuzz message
I had to move Verify function into separate file to tag it as `cgo` only
build. go-fuzz doesn't work with cgo.

License: MIT
